fix: retry HTTP/2 protocol errors
