+++ lychee-lib/src/types/status.rs
@@ -1,23 +1,13 @@
-use std::{collections::HashSet, error::Error, fmt::Display};
+use std::{collections::HashSet, fmt::Display};
 
 use super::CacheStatus;
 use crate::ErrorKind;
 use crate::RequestError;
 use crate::ratelimit::CacheableResponse;
+use crate::retry::is_http2_error;
 use http::StatusCode;
 use serde::ser::SerializeStruct;
 use serde::{Serialize, Serializer};
-
-fn is_http2_error(error: &(dyn Error + 'static)) -> bool {
-    let mut current = Some(error);
-    while let Some(error) = current {
-        if error.downcast_ref::<h2::Error>().is_some() {
-            return true;
-        }
-        current = error.source();
-    }
-    false
-}
 
 const ICON_OK: &str = "✔";
 const ICON_EXCLUDED: &str = "?";
@@ -319,7 +309,6 @@ impl From<ErrorKind> for Status {
 
 #[cfg(test)]
 mod tests {
-    use super::is_http2_error;
     use crate::{CacheStatus, ErrorKind, Status};
     use http::StatusCode;
 
@@ -389,17 +378,5 @@ mod tests {
     fn test_status_unknown() {
         assert!(Status::UnknownStatusCode(StatusCode::from_u16(999).unwrap()).is_unknown());
         assert!(!Status::Ok(StatusCode::from_u16(200).unwrap()).is_unknown());
-    }
-
-    #[test]
-    fn identifies_http2_protocol_errors_by_type() {
-        let error = h2::Error::from(h2::Reason::PROTOCOL_ERROR);
-        assert!(is_http2_error(&error));
-    }
-
-    #[test]
-    fn ignores_unrelated_network_errors() {
-        let error = std::io::Error::other("connection reset by peer");
-        assert!(!is_http2_error(&error));
     }
 }
